requirements.txt
PyNaCl~=1.5.0
requests~=2.32.3
python-dotenv~=1.0.1
